hide package:image dependency and fix import spelling
