+++ packages/flutter_test/lib/src/_goldens_io.dart
@@ -4,10 +4,11 @@
 
 import 'dart:async';
 import 'dart:io';
+import 'dart:math' as math;
 import 'dart:typed_data';
-import 'dart:ui';
 
 import 'package:image/image.dart';
+import 'package:meta/meta.dart';
 import 'package:path/path.dart' as path;
 import 'package:test_api/test_api.dart' as test_package show TestFailure;
 
@@ -63,7 +64,7 @@ class LocalFileComparator extends GoldenFileComparator {
       throw test_package.TestFailure('Could not be compared against non-existent file: "$golden"');
     }
     final List<int> goldenBytes = await goldenFile.readAsBytes();
-    final ComparisonResult result = GoldenFileComparator.compareLists<Uint8List>(imageBytes, goldenBytes);
+    final ComparisonResult result = compareLists<Uint8List>(imageBytes, goldenBytes);
 
     if (!result.passed) {
       String additionalFeedback = '';
@@ -102,3 +103,102 @@ class LocalFileComparator extends GoldenFileComparator {
   }
 }
 
+/// Returns a [ComparisonResult] to describe the pixel differential of the
+/// [test] and [master] image bytes provided.
+ComparisonResult compareLists<T>(List<int> test, List<int> master) {
+  if (identical(test, master))
+    return ComparisonResult(passed: true);
+
+  if (test == null || master == null || test.isEmpty || master.isEmpty) {
+    return ComparisonResult(
+      passed: false,
+      error: 'Pixel test failed, null image provided.',
+    );
+  }
+
+  final Image testImage = decodePng(test);
+  final Image masterImage = decodePng(master);
+
+  assert(testImage != null);
+  assert(masterImage != null);
+
+  final int width = testImage.width;
+  final int height = testImage.height;
+
+  if (width != masterImage.width || height != masterImage.height) {
+    return ComparisonResult(
+      passed: false,
+      error: 'Pixel test failed, image sizes do not match.\n'
+        'Master Image: ${masterImage.width} X ${masterImage.height}\n'
+        'Test Image: ${testImage.width} X ${testImage.height}',
+    );
+  }
+
+  int pixelDiffCount = 0;
+  final int totalPixels = width * height;
+  final Image invertedMaster = invert(Image.from(masterImage));
+  final Image invertedTest = invert(Image.from(testImage));
+
+  final Map<String, Image> diffs = <String, Image>{
+    'masterImage' : masterImage,
+    'testImage' : testImage,
+    'maskedDiff' : Image.from(testImage),
+    'isolatedDiff' : Image(width, height),
+  };
+
+  for (int x = 0; x < width; x++) {
+    for (int y =0; y < height; y++) {
+      final int testPixel = testImage.getPixel(x, y);
+      final int masterPixel = masterImage.getPixel(x, y);
+
+      final int diffPixel = (getRed(testPixel) - getRed(masterPixel)).abs()
+        + (getGreen(testPixel) - getGreen(masterPixel)).abs()
+        + (getBlue(testPixel) - getBlue(masterPixel)).abs()
+        + (getAlpha(testPixel) - getAlpha(masterPixel)).abs();
+
+      if (diffPixel != 0 ) {
+        final int invertedMasterPixel = invertedMaster.getPixel(x, y);
+        final int invertedTestPixel = invertedTest.getPixel(x, y);
+        final int maskPixel = math.max(invertedMasterPixel, invertedTestPixel);
+        diffs['maskedDiff'].setPixel(x, y, maskPixel);
+        diffs['isolatedDiff'].setPixel(x, y, maskPixel);
+        pixelDiffCount++;
+      }
+    }
+  }
+
+  if (pixelDiffCount > 0) {
+    return ComparisonResult(
+      passed: false,
+      error: 'Pixel test failed, ${((pixelDiffCount/totalPixels) * 100).toStringAsFixed(2)}% diff detected.',
+      diffs: diffs,
+    );
+  }
+  return ComparisonResult(passed: true);
+}
+
+/// The result of a pixel comparison test.
+///
+/// The [ComparisonResult] will always indicate if a test has [passed]. The
+/// optional [error] and [diffs] parameters provide further information about
+/// the result of a failing test.
+class ComparisonResult {
+  /// Creates a new [ComparisonResult] for the current test.
+  ComparisonResult({
+    @required this.passed,
+    this.error,
+    this.diffs,
+  }) : assert(passed != null);
+
+  /// Indicates whether or not a pixel comparison test has failed.
+  ///
+  /// This value cannot be null.
+  final bool passed;
+
+  /// Error message used to describe the cause of the pixel comparison failure.
+  final String error;
+
+  /// Map containing differential images to illustrate found variants in pixel
+  /// values in the execution of the pixel test.
+  final Map<String, Image> diffs;
+}
